+++ packages/react-components/react-dialog/src/components/Dialog/useDialog.ts
@@ -1,10 +1,16 @@
 import * as React from 'react';
-import { useControllableState, useEventCallback, useId, useIsomorphicLayoutEffect } from '@fluentui/react-utilities';
+import {
+  useControllableState,
+  useEventCallback,
+  useId,
+  useIsomorphicLayoutEffect,
+  useMergedRefs,
+} from '@fluentui/react-utilities';
+import { useHasParentContext } from '@fluentui/react-context-selector';
+import { useControlNativeDialogOpenState, useDisableBodyScroll, useFocusFirstElement } from '../../utils';
+import { DialogContext } from '../../contexts';
 
 import type { DialogOpenChangeData, DialogProps, DialogState } from './Dialog.types';
-import { useDisableBodyScroll, useFocusFirstElement } from '../../utils';
-import { useHasParentContext } from '@fluentui/react-context-selector';
-import { DialogContext } from '../../contexts/dialogContext';
 
 /**
  * Create the state required to render Dialog.
@@ -36,6 +42,7 @@ export const useDialog_unstable = (props: DialogProps): DialogState => {
   });
 
   const focusRef = useFocusFirstElement(open, modalType);
+  const nativeControlRef = useControlNativeDialogOpenState(open, modalType);
   const disableBodyScroll = useDisableBodyScroll();
   const isBodyScrollLocked = Boolean(open && modalType !== 'non-modal');
 
@@ -57,7 +64,7 @@ export const useDialog_unstable = (props: DialogProps): DialogState => {
     dialogBodyID: useId('dialog-body-'),
     dialogTitleID: useId('dialog-title-'),
     isNestedDialog: useHasParentContext(DialogContext),
-    dialogRef: focusRef,
+    dialogRef: useMergedRefs(focusRef, nativeControlRef),
   };
 };