+++ packages/react-components/react-table/src/components/TableCell/useTableCellStyles.ts
@@ -6,7 +6,6 @@ import type { SlotClassNames } from '@fluentui/react-utilities';
 export const tableCellClassName = 'fui-TableCell';
 export const tableCellClassNames: SlotClassNames<TableCellSlots> = {
   root: tableCellClassName,
-  media: 'fui-TableCell__media',
 };
 
 /**
@@ -14,16 +13,9 @@ export const tableCellClassNames: SlotClassNames<TableCellSlots> = {
  */
 const useStyles = makeStyles({
   root: {
+    verticalAlign: 'middle',
+    display: 'table-cell',
     ...shorthands.padding('0px', tokens.spacingHorizontalS),
-    display: 'flex',
-    alignItems: 'center',
-    ...shorthands.gap(tokens.spacingHorizontalS),
-    ...shorthands.flex(1, 1, '0px'),
-  },
-
-  media: {
-    display: 'flex',
-    alignItems: 'center',
   },
 });
 
@@ -33,9 +25,5 @@ const useStyles = makeStyles({
 export const useTableCellStyles_unstable = (state: TableCellState): TableCellState => {
   const styles = useStyles();
   state.root.className = mergeClasses(tableCellClassNames.root, styles.root, state.root.className);
-  if (state.media) {
-    state.media.className = mergeClasses(tableCellClassNames.media, styles.media);
-  }
-
   return state;
 };